+++ src/api/ComputeResponseImpl.java
@@ -1,6 +1,8 @@
 package api;
 
-class ComputeResponseImpl implements ComputeResponse {
+import java.util.List;
+
+public class ComputeResponseImpl implements ComputeResult {
     private final boolean success;
     private final String errorMessage;
 
@@ -15,6 +17,12 @@ class ComputeResponseImpl implements ComputeResponse {
     }
 
     @Override
+    public List<Integer> getFactors() {
+        // Since this class represents a response rather than actual computation results,
+        // we can return null or an empty list if success is false
+        return null;
+    }
+
     public String getErrorMessage() {
         return errorMessage;
     }
